sem_5/OIIS/lab_2/requirements.txt
Pillow>=10,<11
pillow-heif>=0.17

